requirements.txt
docx==0.2.4
docx2pdf==0.1.8
Flask==2.1.2
Flask_Login==0.6.1
flask_mongoengine==1.0.0
Flask_WTF==1.0.1
google_api_python_client==2.50.0
oauth2client==4.1.3
pandas==1.1.0
protobuf==4.21.1
pymongo==3.11.0
python-dotenv==0.20.0
python_docx==0.8.11
pywin32==228
requests==2.26.0
Werkzeug==2.1.2
WTForms==3.0.1
